requirements.txt
litserve>=0.2.3  # Adjust this version based on the latest available
torch>=1.13.1     # Required for running FasterWhisper
torchaudio>=0.13.1  # For handling audio files
faster-whisper==0.2.0  # Install the specific version of FasterWhisper
ffmpeg-python     # For handling ffmpeg operations if needed
python-multipart  # For handling file uploads
moviepy==1.0.3  # For handling video files
requests>=2.28.1  # For handling HTTP requests
opencv-python>=4.8.0.76  # For handling image files
fastapi
boto3>=1.26.0  # For R2/S3 bucket operations
python-dotenv>=0.19.0  # For loading .env files